requirements.txt
google-generativeai
slack-sdk
sendgrid
requests
python-dotenv
